#KT-349 - Avropare skall kunna delegera när som helst i flödet, samt ändra delegering: Added function to change/remove/add contactperson of order:
OrderContactPersonHistory new table for history and tracing of who and when a change was done.

Policies - removed code from Edit that was used by Replacing order of  to a separate Replace-handler. Edit-handler for Order is now used for edit a contact person (and will be used for edit other parts such as attachments etc). Removed unused Edit-code for Request.