#KT-601 - UX - Fixa rekvisitionvyer: Fix with view for requisition in new tab-view.
_RequisitionSummary - partial for "new invoice design" used by View and Process.

View.cshtml - included the process-buttons for customer in this view.
Process.cshtml - keep it for now, but will probably be deleted.

NotificationService, PriceCalculationService, DisplayPriceInformation, _PriceInformationDisplay - new look and properties for PriceinformationDisplay.

OrderController, RequestModel - corrected bug and made more light possibility if viewing request details from order.

PreviousRequisitionViewModel - corrected bug with WasteTime

Partial views, site.js - changed collapsable to view a h3-header if not collapsable and else a smaller as a part of another page. Changed the java script but it is still not workoing on tab views.