#KT-604 - Förmedling ska kunna bekräfta att de sett att myndighet avböjt tolktillsättning:
Possibility for broker to confirm denial of request. New table RequestStatusConfirmation for confirmingUser and confirmtime etc. If confirmed - display it in Eventlog and remove from start reportlist.

RequestSummary -  Bug fix, broker should never see orderstatus (cause if first broker doesn't have the active it is totaly wrong, and both broker and customer have order- or requeststatus in header so no need for it here.

Site.less - fix with style for start reportlist.