#KT-950 - Hantera användares enheter från användarsidorna: Functionality to handle customer units from edit user page. Local admins can only handle their local admin units but other units are displayed.
Audit for change of unit user.

For this checkin (since unit pages with user is not developed yet and a change) local admins have their own user menu and are authorized to view/edit all users of the customer. This will be changed in later checkin.

Index.cshtml - changed table to more consistent column names and column order.