#KT-987 - Användarhantering/policies för lokala admins: Fix with policies and user handling for local admins.
Local admins must not edit users that are not connected to their local admin units.

Local admins should only reach users from units menu (no users menu).

Local admins must not change if a user should be active or not (but display if user is Active and OrganisationAdministrator on edit page for local admins).

Users.cshtml - improvements of UI.

Renamed IsCentralAdministrator to IsOrganisationAdministrator.