#KT-1009 - Varna beställaren om hen försöker beställa en tolk med en kompetens som ingen tolk i sverige har:
New integration to Tellus API to get info if interpreters exist for the competence levels for a language.

Warn and display information to customer if there are no interpreters for required competence level - Information is displayed in differnt colors when competence levels are selected (page one) on page two (confirm) the user is warned if there is at least one competence level that is required but not exist for language.

Made better language pages for application administrators (all language actions/views are moved to LanguageController). Improvements of list languages.