Updated nuget packages as first step in new version.
Had to refactor the use of AutoMapper since the new version of AutoMapper did not support the way the map(OrderModel => ReplacedOrderModel) was initialized.
Now the AutoMapper setup is more in tune with .net core and stuff. This also removed a warning about the previous implementation, that warned about the upcoming deprecation.