KT-327 - Avropare vill kunna skriva ut bekräftelsen
Added possibility to print an approved request for customer.
Added link to print request in emails (if approved).
Print includes the most important info from order and the info in request answer.
Added possibility to display ErrorMessage for Order/View to display that request has wrong status to be printed (for examle if link in email is used after a request is cancelled).