#KT-1238 - Godkänna tillsättning - tydligare: Fix when customer shall respond a request answer.
When customer approves a group get the cost and info to cost more visible if no extra interpreter.

Customer must not respond an individual order in Order/View - display info to go to group.

Should be possible to accept an interpreter change in Order/View.

Broker must not change interpreter until the group is approved.

Customer must not cancel order until the group is approved.