#KT-1252 - Bättre hantering av skyddade tolkar: Improve handling of protected interpreters
Use a flag in database for proteced interpreter.
Broker should not be able to handle/edit the interpreters in interpreter list.
Do not validate against Tellus if assigning protected interpreters (should not have a KamK Official Id).
Script for setting data for the protected interpreters.

#KT-1222 - Det skall inte gå att lägga till samma tolk när man tillsätter två tolkar (extra tolk):
Shouldn't be possible to assign same interpreter when two interpreters should be assigned - Fix part 2 - two protected interpreters should be possible to assign.