#KT-1013 - Gör det möjligt för en förmedlare att hantera de beställningar som inväntar rekvisition: Possible to handle requests where order has started/awaiting requisition:
It should be possible to archive a request without doing a requisition so it dissapears from startlist - ConfirmNoRequisition (also from API).

New requeststatus Delivered that is set when archive without requisition and when doing requisition. Handle the new status in different places.
Script for setting new status for requests that have requisitions.
Display ConfirmNoRequisition in eventlog.

Tests for ConfirmNoRequisition and handle of requeststatus Delivered.

HomeController - always go to Request/View with the two choices when handle an approved request where order started.