Fix with Accept for broker (and some bugs on the way): (All is not working correctly, more fixes to come -AllowExceedingTravelCost for broker, process.js does not work with LatestAnswerTime, Replacement link for broker is wrong)
Improvements of SQL in Accept for RequestController

OrderController - View and Print were crashing if ExpectedTravelCosts should be dispalyed.

OrderViewModel, _OrderSummary- IsCreatorInterpreterUser was always null (Ej angivet).

RequestSummary, RequestModel, Process.cshtml - when replacement order should be processed it was a crash due to new OrderViewModel.

_RequestSummary, RequestViewModel - A lot of things were not displayed when processing a replacement order since IsReplacedRequest was always false  And a merge-fix of css-class.

ListToModelService - if language was not an authorized language the requested competence level was dispalyed as 1. Ej tillsatt/not assigned instead of the info text that language lasks authorization.

View.cshtml (Order and Request) - non answeered requests displayed a zero cost PriceInformation -removed that.