#KT-1402 - Ange bokningsnummer vanlig förfrågan:
Add referencenumber for request for broker:

Posiible to set when answering request and when changing interpreter.
Display for broker, sysadmin and appadmin
Add to notification emails for broker.
Add in API when answering request and view for request
Add in webhook (useful when a replacement order is created).

Startup.cs, IModel.cs, Ordergroup.cs - fix build meassages.